refactor(app): 拆分 ImageUpload 子组件
提取 Dropzone、确认表单、压缩面板、批量进度与工具函数，主文件降至约 1070 行。
